Rename library to async
