+++ .flake8
@@ -8,7 +8,7 @@ ignore = E203,E501,E731,W503,W605
 import-order-style = google
 # Packages added in this list should be added to the setup.cfg file as well
 application-import-names =
-    cubejsclient
+    cubejsclientasync
 exclude =
     *vendor*
     .venv
