+++ setup.cfg
@@ -10,7 +10,7 @@ force_sort_within_sections=true
 include_trailing_comma=true
 known_standard_library=typing
 known_first_party=
-    cubejsclient
+    cubejsclientasync
 line_length=88
 multi_line_output=3
 no_lines_before=LOCALFOLDER
